+++ serial_bitCount.cpp
@@ -1,7 +1,7 @@
 #include "stdint.h"
 
-int naive_countOnes(uint64_t c) {
-    int count = 0;
+uint64_t naive_countOnes(uint64_t c) {
+    uint64_t count = 0;
     while (c) {
         count += c & 1;
         c = c >> 1;
@@ -9,12 +9,12 @@ int naive_countOnes(uint64_t c) {
     return count;
 }
 
-int naive_popcount(uint64_t c) {
+uint64_t naive_popcount(uint64_t c) {
     return __builtin_popcountll(c);
 }
 
-int countOnes(uint64_t* buffer, int size) {
-    int count = 0;
+uint64_t countOnes(uint64_t* buffer, int size) {
+    uint64_t count = 0;
     for (int i = 0; i < size; i++) {
         //count += naive_countOnes(buffer[i]);
         count += naive_popcount(buffer[i]);
